Get rid of all the setup functions in the waf tools
The old model (copied from Scons) was to store logic (mapping file extensions to functions)
along with the data. In Waf a way was found to separate that logic by adding an indirection
layer (storing the names in the Environment instances)

In the new model, the logic is more object-oriented, and the user scripts provide the
logic. The data files (Environments) must contain configuration data only (flags, ..).



git-svn-id: http://waf.googlecode.com/svn/trunk@2514 f0382ac9-c320-0410-b3f0-b508d59f5a85