Make demos/mac_app/ more like a real Mac App.
The sources and resources have been moved to subdirectories, so named.  This
necessitates a change to the processing of the 'mac_resources' attribute,
whereby only the basename of the provided resources will be used, and whereby
directories provided there will be expanded recursively.

Logically, a nib bundle is a single unit; a single file from a nib bundle would
never be included independently of any other.  So, specifying the path to the
bundle matches the mental model of the developer better, while not leading to
any real loss of expressiveness.

git-svn-id: http://waf.googlecode.com/svn/trunk@11202 f0382ac9-c320-0410-b3f0-b508d59f5a85